Vlookup project 2
This is an MIS report prepared with the help of Vlookup tool of excel. With the help of this report, one can easily find out the name of the product sold by the particular salesmen from a list of 32 salesmen. One can also find the quantity of the items sold, the unit price of the item, and the total amount earned by that particular salesman, at the same click.